requirements.txt
# PyTorch深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# TensorFlow Lite (用于face_blendshapes模型)
tensorflow>=2.6.0

# 数据处理
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# 配置文件
PyYAML>=5.4.0

# 进度条
tqdm>=4.62.0

# 可视化 (可选)
matplotlib>=3.4.0
seaborn>=0.11.0

# 日志和工具
tensorboard>=2.7.0  # 可选，用于训练监控

# 开发工具 (可选)
jupyter>=1.0.0
notebook>=6.4.0

# 性能优化 (可选)
psutil>=5.8.0  # 系统监控
